Add TS to netlify func
