+++ netlify.toml
@@ -1,2 +1,4 @@
 [dev]
   framework = "vite"
+[functions]
+  included_files = ["./netlify/functions/writeToSpreadsheet/assets/**"]
