+++ netlify/functions/writeToSpreadsheet/assets/logga.png
