+++ netlify/functions/writeToSpreadsheet/writeToSpreadsheet.ts
@@ -1,16 +1,21 @@
 if (!process.env.NETLIFY) {
   require("dotenv").config();
 }
-const moment = require("moment-timezone");
-const { JWT } = require('google-auth-library');
-const axios = require("axios");
+import type { Handler, HandlerEvent, HandlerContext } from "@netlify/functions";
+import moment from "moment-timezone";
+import { JWT } from 'google-auth-library';
+import axios from "axios";
+import { FormType } from "../../../src/components/register/models";
+import { appendZero } from "../../../src/Utils";
+import { sendEmail } from "./emailSender";
+import { GoogleSpreadsheet } from "google-spreadsheet";
 
 // required env vars
 if (!process.env.GOOGLE_SERVICE_ACCOUNT_EMAIL)
   throw new Error("no GOOGLE_SERVICE_ACCOUNT_EMAIL env var set");
 if (!process.env.GOOGLE_PRIVATE_KEY)
   throw new Error("no GOOGLE_PRIVATE_KEY env var set");
-  // spreadsheet key is the long id in the sheets URL
+// spreadsheet key is the long id in the sheets URL
 if (!process.env.GOOGLE_SPREADSHEET_ID_SOLO_2024)
   throw new Error("no GOOGLE_SPREADSHEET_ID_SOLO_2024 env var set");
 if (!process.env.GOOGLE_SPREADSHEET_ID_TEAM_2024)
@@ -18,14 +23,7 @@ if (!process.env.GOOGLE_SPREADSHEET_ID_TEAM_2024)
 if (!process.env.GOOGLE_SPREADSHEET_ID_TSHIRT_ORDER)
   throw new Error("no GOOGLE_SPREADSHEET_ID_TSHIRT_ORDER env var set");
 
-const { GoogleSpreadsheet } = require("google-spreadsheet");
-const sendEmail = require("./emailSender");
-
-function appendZero(str) {
-  return parseInt(str) < 10 ? "0" + str : str;
-}
-
-function handleBirthday(data, type) {
+function handleBirthday(data: any, type: FormType) {
   if (type == "team") {
     data["birthday1"] =
       data["year1"] + "-" + appendZero(data["month1"]) + "-" + appendZero(data["day1"]);
@@ -40,66 +38,82 @@ function handleBirthday(data, type) {
   return data;
 }
 
-exports.handler = async (event, context, callback) => {
+const handler: Handler = async (event: HandlerEvent, context: HandlerContext) => {
   // Check recaptcha from token
   try {
-    const result = await axios.post(`https://www.google.com/recaptcha/api/siteverify?secret=${process.env.VITE_RECAPTCHA_SECRET}&response=${event.queryStringParameters.token}`);
+    const result = await axios.post(`https://www.google.com/recaptcha/api/siteverify?secret=${process.env.VITE_RECAPTCHA_SECRET}&response=${event.queryStringParameters?.token}`);
     if (result.data.score < 0.5) {
       return {
         statusCode: 400,
-          body: JSON.stringify({
-            message: "Är du en robot?",
-          }),
+        body: JSON.stringify({
+          message: "Are you a robot?",
+        }),
       }
     }
   } catch (e) {
     return {
       statusCode: 500,
-        body: JSON.stringify({
-          message: "Kunde inte processera förfrågan.",
-        }),
+      body: JSON.stringify({
+        message: "Could not verify captcha.",
+      }),
     }
   }
 
   let spreadsheetID = "";
-  let idType  = "";
+  let idType = "";
+
   console.log("Running sheet netlify function...");
 
-  const registerType = event.queryStringParameters.type;
+  const registerTypeParam = event.queryStringParameters?.type;
+  let registerType: FormType = "solo";
 
-  switch (registerType) {
+  switch (registerTypeParam) {
     case "solo":
-      spreadsheetID = process.env.GOOGLE_SPREADSHEET_ID_SOLO_2024;
+      spreadsheetID = process.env.GOOGLE_SPREADSHEET_ID_SOLO_2024!;
       idType = "S";
       break;
     case "team":
-      spreadsheetID = process.env.GOOGLE_SPREADSHEET_ID_TEAM_2024;
+      spreadsheetID = process.env.GOOGLE_SPREADSHEET_ID_TEAM_2024!;
       idType = "T";
+      registerType = "team";
       break;
     case "tshirt_order":
-      spreadsheetID = process.env.GOOGLE_SPREADSHEET_ID_TSHIRT_ORDER;
+      spreadsheetID = process.env.GOOGLE_SPREADSHEET_ID_TSHIRT_ORDER!;
       idType = "O";
+      registerType = "tshirt_order";
       break;
     default:
-      console.log("The type given is not valid!");
-      break;
+      return {
+        statusCode: 400,
+        body: "No valid registration type",
+      };
   }
 
   try {
     const serviceAccountAuth = new JWT({
       email: process.env.GOOGLE_SERVICE_ACCOUNT_EMAIL,
-      key: process.env.GOOGLE_PRIVATE_KEY.replace(/\\n/g, "\n"),
+      key: process.env.GOOGLE_PRIVATE_KEY?.replace(/\\n/g, "\n"),
       scopes: [
         'https://www.googleapis.com/auth/spreadsheets',
       ],
     });
 
     const doc = new GoogleSpreadsheet(spreadsheetID, serviceAccountAuth);
-    
+
     await doc.loadInfo();
     const sheet = doc.sheetsByIndex[0];
 
-    let data = JSON.parse(event.body);
+    // TODO: Handle types
+    let parsedData: any = null;
+
+    try {
+      parsedData = JSON.parse(event.body ?? "");
+    } catch (error) {
+      return {
+        statusCode: 400,
+        body: "Invalid JSON",
+      };
+    }
 
     const rows = await sheet.getRows();
     const lastRow = rows[rows.length - 1];
@@ -107,18 +121,20 @@ exports.handler = async (event, context, callback) => {
     // Get the number only and not the letter
     const idNumber = lastRow ? parseInt(lastRow.get('id').substring(1)) : 0;
 
-    data = handleBirthday(data, registerType);
+    parsedData = handleBirthday(parsedData, registerType);
 
-    data["id"] = idType + (idNumber + 1).toString();
-    data["uploadTime"] = moment()
+    parsedData["id"] = idType + (idNumber + 1).toString();
+    parsedData["uploadTime"] = moment()
       .tz("Europe/Stockholm")
       .format("YYYY-MM-DD HH:mm");
 
-    const addedRow = await sheet.addRow(data);
+    console.log("parsedData", parsedData)
+
+    const addedRow = await sheet.addRow(parsedData);
 
     if (addedRow) {
       console.log("Success adding row");
-      const email_sent = await sendEmail(addedRow, registerType);
+      const email_sent = await sendEmail(addedRow, registerTypeParam);
       return {
         statusCode: 200,
         body: JSON.stringify({
@@ -134,11 +150,15 @@ exports.handler = async (event, context, callback) => {
         }),
       };
     }
-  } catch (e) {
+  } catch (e: unknown) {
     console.error(e)
     return {
       statusCode: 500,
-      body: e.toString(),
+      body: JSON.stringify({
+        message: `An unknown error occured`,
+      }),
     };
   }
 };
+
+export { handler };
