+++ package.json
@@ -49,6 +49,7 @@
     ]
   },
   "devDependencies": {
+    "@types/nodemailer": "^6.4.16",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@types/react-image-gallery": "^1.2.4",
