+++ src/components/register/models.ts
@@ -4,16 +4,65 @@ export interface Shirt {
   material: "funktion" | "bomull"
 }
 
-export interface orderDetails {
-  loading: boolean,
+export interface BaseOrderType {
+  shirts: Shirt[],
+  shirtsString?: string,
+  numCaps: number,
+  extraDonation: number,
   name: string,
   email: string,
-  extraDonation: number,
-  shirts: Shirt[],
-  numCaps: number,
   info: string,
-  consent: boolean,
-  hasOrderd: boolean
 }
+
+export interface RegisterFormSoloState extends BaseOrderType{
+  year: string;
+  month: string;
+  day: string;
+  gender: string;
+  city: string;
+  isCheckboxOneTicked: boolean;
+  isCheckboxTwoTicked: boolean;
+  isCheckboxThreeTicked: boolean;
+}
+
+export interface RegisterFormTeamState extends Omit<BaseOrderType, "name" | "email">{
+  teamName: string;
+  name1: string;
+  email1: string;
+  year1: string;
+  month1: string;
+  day1: string;
+  city1: string;
+  name2: string;
+  email2: string;
+  year2: string;
+  month2: string;
+  day2: string;
+  city2: string;
+  name3: string;
+  email3: string;
+  year3: string;
+  month3: string;
+  day3: string;
+  city3: string;
+  isCheckboxOneTicked: boolean;
+  isCheckboxTwoTicked: boolean;
+  isCheckboxThreeTicked: boolean;
+}
+
+export interface OrderShirtState extends BaseOrderType {
+  consent: boolean,
+  hasOrdered: boolean,
+  loading: boolean,
+}
+
+export type DataToSend = Partial<
+  BaseOrderType &
+  RegisterFormSoloState &
+  RegisterFormTeamState &
+  OrderShirtState
+> & {
+  totalToPay: number
+};
 
 export type FormType = "solo" | "team" | "tshirt_order";
